+++ library/src/test/java/com/twilio/video/ConnectOptionsUnitTest.java
@@ -2,6 +2,7 @@ package com.twilio.video;
 
 import org.junit.Test;
 
+import static junit.framework.TestCase.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 public class ConnectOptionsUnitTest {
@@ -15,14 +16,27 @@ public class ConnectOptionsUnitTest {
         new ConnectOptions.Builder("").build();
     }
 
-    /*
-     * TODO: GSDK-1141 Disabled insights for GA. Uncomment once they are enabled again.
-     */
-//    @Test
-//    public void shouldEnableInsights() {
-//        ConnectOptions connectOptions = new ConnectOptions.Builder("test")
-//                .enableInsights(true)
-//                .build();
-//        assertTrue(connectOptions.isInsightsEnabled());
-//    }
+    @Test
+    public void insightsShouldBeEnabledByDefault() {
+        ConnectOptions connectOptions = new ConnectOptions.Builder("test")
+                .build();
+
+        assertTrue(connectOptions.isInsightsEnabled());
+    }
+
+    @Test
+    public void shouldEnableInsights() {
+        ConnectOptions connectOptions = new ConnectOptions.Builder("test")
+                .enableInsights(true)
+                .build();
+        assertTrue(connectOptions.isInsightsEnabled());
+    }
+
+    @Test
+    public void shouldDisableInsights() {
+        ConnectOptions connectOptions = new ConnectOptions.Builder("test")
+                .enableInsights(false)
+                .build();
+        assertFalse(connectOptions.isInsightsEnabled());
+    }
 }